refactored to support PSR-4 & IronCore 1.0.*
+ moved classes to /src folder
+ moved test into /test folder
+ removed phar & build.php
