+++ .gitignore
@@ -1,3 +1,1 @@
 vendor
-config.ini
-iron.json
